use beta distribution
This has a nice curve which still asks some of the questions which
have good previous scores, although not very often. It is heavily
weighted to those with poor scores.

The value 0.5 provides a fairly heavy weighting. A value of 1 can
be used to spread things out a little more evenly.